refactor: moved callbacks into utilities modules
moved callbacks into the utilities module as its own submodule, callbacks. This change allows me to import from the module and remove the py file from the root directory where drive.py is
